:zap: Avoid Decoding Entire TIFF Before Conversion (#48)
* :recycle: Refactor thumbnail To Use get_tile

* :poop: Fall Back To __getitem__ if get_tile Fails

* :construction: Swap To tifffile ZarrStore

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* :bug: Handle `tifffile.aszarr` When Group Returned

* :recycle: Mark Extra TIFFReader Attrs Private

* :heavy_plus_sign: Add dask For Axis Reordering

* :recycle: Rename Public/Private Variables

* :recycle: Refactor _can_transcode_from_reader

* :heavy_plus_sign: Add pytest-timeout For CI Debugging

* :label: Add Test Type Hints

* :white_check_mark: Set Convert Test Timeout To 25s

* :wrench: Add PyTest norecursedirs For Zarr

* :poo: Fix TIFFReader Dask Array Read From Subprocess

* :heavy_plus_sign: Add Dask Distributed

* :wrench: Increase PyTest Timeout + Add Duration Report

* :wrench: Fix PyTest Opts

* :poo: Use Xarray

* :white_check_mark: Update Thumbnail Test

* :rotating_light: Fix Linter Issues

* :rotating_light: Remove Commented Out Code

* :rotating_light: Add Linter Exception

* :recycle: Refactor Tile Iterator + Add Dask Compatible Iterator

* :fire: Remove Unused Method

* :wrench: Update Pre-Commit Config

* :white_check_mark: Set Compression Level In Test

* 🚨 Add Warning Stack Level

---------
