+++ requirements.txt
@@ -1,6 +1,8 @@
 black==22.10.0
 bump2version==1.0.1
 Click>=8.0.0
+dask
+dask[distributed]
 flake8==6.0.0
 glymur>=0.11.3
 imagecodecs>=2022.7.27
@@ -11,6 +13,7 @@ openslide-python>=1.1.2
 pip==22.2.2
 pytest==7.2.2
 pytest-cov>=3.0.0
+pytest-timeout
 scipy>=1.7.3
 tifffile>=2022.4.28
 tox==4.4.7
@@ -18,4 +21,5 @@ twine==1.14.0
 watchdog==2.3.1
 wheel==0.38.4
 wsidicom>=0.2.0
+xarray
 zarr==2.13.3
