+++ setup.cfg
@@ -27,7 +27,9 @@ per-file-ignores =
 max-cognitive-complexity = 10
 
 [tool:pytest]
-addopts = --cov-config=setup.cfg --ignore=setup.py --cov=wsic --cov-report=term --cov-report=xml:coverage.xml --cov-report=html
+addopts = --cov-config=setup.cfg --ignore=setup.py --cov=wsic --cov-report=term --cov-report=xml:coverage.xml --cov-report=html --durations=10 --durations-min=10
+timeout = 60
+norecursedirs = *.zarr
 
 [tool:isort]
 profile = black
