+++ wsic/readers.py
@@ -1,22 +1,18 @@
 import multiprocessing
-import os
-import time
 import warnings
 from abc import ABC, abstractmethod
-from concurrent.futures import ThreadPoolExecutor
 from contextlib import suppress
-from math import ceil, floor
 from pathlib import Path
-from typing import Dict, Iterable, Iterator, Optional, Tuple, Union
+from typing import Iterable, Iterator, Optional, Tuple, Union
 
 import numpy as np
+import xarray as xr
 import zarr
 
 from wsic.codecs import register_codecs
 from wsic.enums import Codec, ColorSpace
 from wsic.magic import summon_file_types
 from wsic.metadata import ngff
-from wsic.multiproc import Queue
 from wsic.typedefs import PathLike
 from wsic.utils import (
     block_downsample_shape,
@@ -27,7 +23,6 @@ from wsic.utils import (
     scale_to_fit,
     tile_slices,
     warn_unused,
-    wrap_index,
 )
 
 
@@ -95,11 +90,60 @@ class Reader(ABC):
         yx_tile_shape = (
             self.tile_shape[:2] if self.tile_shape else np.minimum(yx_shape, (256, 256))
         )
-        self_mosaic_shape = (
-            self.mosaic_shape
-            if self.mosaic_shape
-            else mosaic_shape(yx_shape, yx_tile_shape)
-        )
+        self_mosaic_shape = self.mosaic_shape or mosaic_shape(yx_shape, yx_tile_shape)
+        (
+            downsample_shape,
+            downsample_tile_shape,
+            downsample,
+        ) = self._find_thumbnail_downsample(shape, yx_shape, yx_tile_shape)
+        # NOTE: Assuming channels last
+        channels = self.shape[-1]
+        thumbnail = np.zeros(downsample_shape + (channels,), dtype=np.uint8)
+        # Resize tiles to new_downsample_tile_shape and combine
+        tile_indexes = list(np.ndindex(self_mosaic_shape))
+        for tile_index in self.pbar(tile_indexes, desc="Generating thumbnail"):
+            try:
+                tile = self.get_tile(tile_index)
+            except (ValueError, NotImplementedError):  # e.g. Not tiled
+                tile = self[tile_slices(tile_index, yx_tile_shape)]
+            tile = mean_pool(tile.astype(float), downsample).astype(np.uint8)
+            # Make sure the tile being written will not exceed the
+            # bounds of the thumbnail
+            yx_position = tuple(
+                i * size for i, size in zip(tile_index, downsample_tile_shape)
+            )
+            max_y, max_x = (
+                min(tile_max, thumb_max - position)
+                for tile_max, thumb_max, position in zip(
+                    tile.shape, thumbnail.shape, yx_position
+                )
+            )
+            sub_tile = tile[:max_y, :max_x]
+            thumbnail[tile_slices(tile_index, downsample_tile_shape)] = sub_tile
+        return thumbnail if approx_ok else resize_array(thumbnail, shape, "bicubic")
+
+    @staticmethod
+    def _find_thumbnail_downsample(
+        thumbnail_shape: Tuple[int, int],
+        yx_shape: Tuple[int, int],
+        yx_tile_shape: Tuple[int, int],
+    ) -> Tuple[Tuple[int, int], Tuple[int, int], int]:
+        """Find the downsample and tile shape for a thumbnail.
+
+        Args:
+            thumbnail_shape (Tuple[int, int]):
+                Shape of the thumbnail to be generated.
+            yx_shape (Tuple[int, int]):
+                Shape of the image in Y and X.
+            yx_tile_shape (Tuple[int, int]):
+                Shape of the tiles in Y and X which will be used to
+                generate the thumbnail.
+
+        Returns:
+            Tuple[Tuple[int, int], Tuple[int, int], int]:
+                Shape of the downsampled image, shape of the downsampled
+                tiles, and the downsample factor.
+        """
         downsample_shape = yx_shape
         downsample_tile_shape = yx_tile_shape
         downsample = 0
@@ -108,23 +152,12 @@ class Reader(ABC):
             next_downsample_shape, new_downsample_tile_shape = block_downsample_shape(
                 yx_shape, new_downsample, yx_tile_shape
             )
-            if not any(x > max(0, y) for x, y in zip(downsample_shape, shape)):
+            if all(x <= max(0, y) for x, y in zip(downsample_shape, thumbnail_shape)):
                 break
             downsample_shape = next_downsample_shape
             downsample_tile_shape = new_downsample_tile_shape
             downsample = new_downsample
-        # NOTE: Assuming channels last
-        channels = self.shape[-1]
-        thumbnail = np.zeros(downsample_shape + (channels,), dtype=np.uint8)
-        # Resize tiles to new_downsample_tile_shape and combine
-        tile_indexes = list(np.ndindex(self_mosaic_shape))
-        for tile_index in self.pbar(tile_indexes, desc="Generating thumbnail"):
-            tile = self[tile_slices(tile_index, yx_tile_shape)]
-            tile = mean_pool(tile.astype(float), downsample).astype(np.uint8)
-            thumbnail[tile_slices(tile_index, downsample_tile_shape)] = tile
-        if approx_ok:
-            return thumbnail
-        return resize_array(thumbnail, shape, "bicubic")
+        return downsample_shape, downsample_tile_shape, downsample
 
     def get_tile(self, index: Tuple[int, int], decode: bool = True) -> np.ndarray:
         """Get tile at index.
@@ -178,337 +211,10 @@ class Reader(ABC):
 
         return tqdm(iterable, *args, **kwargs)
 
-
-def get_tile(
-    queue: Queue,
-    ji: Tuple[int, int],
-    tilesize: Tuple[int, int],
-    path: Path,
-) -> np.ndarray:
-    """Append a tile read from a reader to a multiprocessing queue.
-
-    Args:
-        queue (Queue):
-            A multiprocessing Queue to put tiles on to.
-        ji (Tuple[int, int]):
-            Index of tile.
-        tilesize (Tuple[int, int]):
-            Tile size as (width, height).
-        path (Path):
-            Path to file to read from.
-
-    Returns:
-        Tuple[Tuple[int, int], np.ndarray]:
-            Tuple of the tile index and the tile.
-    """
-    reader = Reader.from_file(path)
-    # Read the tile
-    j, i = ji
-    index = (
-        slice(j * tilesize[1], (j + 1) * tilesize[1]),
-        slice(i * tilesize[0], (i + 1) * tilesize[0]),
-    )
-    # Filter warnings (e.g. from gylmur about reading past the edge)
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        tile = reader[index]
-    queue.put((ji, tile))
-
-
-class MultiProcessTileIterator:
-    """An iterator which returns tiles generated by a reader.
-
-    This is a fancy iterator that uses a multiprocress queue to
-    accelerate the reading of tiles. It can also use an intermediate
-    file to allow for reading and writing with different tile sizes.
-
-    Args:
-        reader (Reader):
-            Reader for image.
-        read_tile_size (Tuple[int, int]):
-            Tile size to read from reader.
-        yield_tile_size (Optional[Tuple[int, int]]):
-            Tile size to yield. If None, yield_tile_size = read_tile_size.
-        num_workers (int):
-            Number of workers to use.
-        intermediate (Path):
-            Intermediate reader/writer to use. Must support random
-            access reads and writes.
-        verbose (bool):
-            Verbose output.
-
-    Yields:
-        np.ndarray:
-            A tile from the reader.
-    """
-
-    def __init__(
-        self,
-        reader: Reader,
-        read_tile_size: Tuple[int, int],
-        yield_tile_size: Optional[Tuple[int, int]] = None,
-        num_workers: int = None,
-        intermediate=None,
-        verbose: bool = False,
-        timeout: float = 10.0,
-        match_tile_sizes: bool = True,
-    ) -> None:
-        self.reader = reader
-        self.shape = reader.shape
-        self.read_tile_size = read_tile_size
-        self.yield_tile_size = yield_tile_size or read_tile_size
-        self.intermediate = intermediate
-        self.verbose = verbose
-        self.timeout = timeout if timeout >= 0 else float("inf")
-        self.processes: Dict[Tuple[int, ...], multiprocessing.Process] = {}
-        self.queue = Queue()
-        self.enqueued = set()
-        self.reordering_dict = {}
-        self.read_j = 0
-        self.read_i = 0
-        self.yield_i = 0
-        self.yield_j = 0
-        self.num_workers = num_workers or os.cpu_count() or 2
-        self.read_mosaic_shape = mosaic_shape(
-            self.shape,
-            self.read_tile_size[::-1],
-        )
-        self.yield_mosaic_shape = mosaic_shape(
-            self.shape,
-            self.yield_tile_size[::-1],
-        )
-        self.remaining_reads = list(np.ndindex(self.read_mosaic_shape))
-        self.tile_status = zarr.zeros(self.yield_mosaic_shape, dtype="u1")
-        try:
-            from tqdm.auto import tqdm
-
-            self.read_pbar = tqdm(
-                total=np.prod(self.read_mosaic_shape),
-                desc="Reading",
-                colour="cyan",
-                smoothing=0.01,
-            )
-        except ImportError:
-            self.read_pbar = None
-
-        # Validation and error handling
-        read_matches_yield = self.read_tile_size == self.yield_tile_size
-        if match_tile_sizes and not read_matches_yield and not self.intermediate:
-            raise ValueError(
-                f"read_tile_size ({self.read_tile_size})"
-                f" != yield_tile_size ({self.yield_tile_size})"
-                " and intermediate is not set. An intermediate is"
-                " required when the read and yield tile size differ."
-            )
-
-    def __len__(self) -> int:
-        """Return the number of tiles in the reader."""
-        return int(np.prod(self.yield_mosaic_shape))
-
-    def __iter__(self) -> Iterator:
-        """Return an iterator for the reader."""
-        self.read_j = 0
-        self.read_i = 0
-        return self
-
     @property
-    def read_index(self) -> Tuple[int, int]:
-        """Return the current read index."""
-        return self.read_j, self.read_i
-
-    @read_index.setter
-    def read_index(self, value: Tuple[int, int]) -> None:
-        """Set the current read index."""
-        self.read_j, self.read_i = value
-
-    @property
-    def yield_index(self) -> Tuple[int, int]:
-        """Return the current yield index."""
-        return self.yield_j, self.yield_i
-
-    @yield_index.setter
-    def yield_index(self, value: Tuple[int, int]) -> None:
-        """Set the current yield index."""
-        self.yield_j, self.yield_i = value
-
-    def wrap_indexes(self) -> None:
-        """Wrap the read and yield indexes."""
-        self.read_index, overflow = wrap_index(self.read_index, self.read_mosaic_shape)
-        if overflow and self.verbose:
-            print("All tiles read.")
-        self.yield_index, overflow = wrap_index(
-            self.yield_index, self.yield_mosaic_shape
-        )
-        if overflow:
-            if self.verbose:
-                print("All tiles yielded.")
-            self.close()
-            raise StopIteration
-
-    def __next__(self) -> np.ndarray:
-        """Return the next tile from the reader."""
-        # Ensure a valid read ij index
-        self.wrap_indexes()
-
-        # Add tile reads to the queue until the maximum number of
-        # workers is reached
-        self.fill_queue()
-
-        # Get the next yield tile from the queue
-        t0 = time.perf_counter()
-        while (time.perf_counter() - t0) < self.timeout:
-            # Remove all tiles from the queue into the reordering dict
-            self.empty_queue()
-
-            # Remove the next read tile from the reordering dict. May be
-            # None if the read tile is not in the reordering dict or if
-            # an intermediate is used.
-            tile = self.pop_next_read_tile()
-
-            # Return the tile if no intermediate is being used and the
-            # tile was in the reordering dict.
-            if not self.intermediate and tile is not None:
-                return tile
-
-            # Get the next tile from the intermediate. Returns None if
-            # intermediate is None or the tile is not in the
-            # intermediate.
-            tile = self.read_next_from_intermediate()
-            if tile is not None:
-                return tile
-
-            # Ensure the queue is kept full
-            self.fill_queue()
-
-            # Sleep and try again
-            time.sleep(0.1)
-        warnings.warn(
-            "Failed to get next tile after 100 attempts. Dumping debug information.",
-            stacklevel=2,
-        )
-        print(f"Reader Shape {self.reader.shape}")
-        print(f"Read Tile Size {self.read_tile_size}")
-        print(f"Yield Tile Size {self.yield_tile_size}")
-        print(f"Read Mosaic Shape {self.read_mosaic_shape}")
-        print(f"Yield Mosaic Shape {self.yield_mosaic_shape}")
-        print(f"Read Index {self.read_index}")
-        print(f"Yield Index {self.yield_index}")
-        print(f"Remaining Reads (:10) {self.remaining_reads[:10]}")
-        print(f"Enqueued {self.enqueued}")
-        print(f"Reordering Dict (keys) {self.reordering_dict.keys()}")
-        print(f"Queue Size {self.queue.qsize()}")
-        intermediate_read_slices = tile_slices(
-            index=(self.yield_j, self.yield_i),
-            shape=self.yield_tile_size[::-1],
-        )
-        print(f"Intermediate Read slices {intermediate_read_slices}")
-        # Terminate the read processes
-        self.close()
-        raise IOError(f"Tile read timed out at index {self.yield_index}")
-
-    def read_next_from_intermediate(self) -> Optional[np.ndarray]:
-        """Read the next tile from the intermediate file."""
-        if self.intermediate is None:
-            return None
-        intermediate_read_slices = tile_slices(
-            index=(self.yield_j, self.yield_i),
-            shape=self.yield_tile_size[::-1],
-        )
-        status = self.tile_status[self.yield_index]
-        if np.all(status == 1):  # Intermediate has all data for the tile
-            self.tile_status[self.yield_index] = 2
-            self.yield_i += 1
-            return self.intermediate[intermediate_read_slices]
-        return None
-
-    def empty_queue(self) -> None:
-        """Remove all tiles from the queue into the reordering dict."""
-        while not self.queue.empty():
-            ji, tile = self.queue.get()
-            self.reordering_dict[ji] = tile
-            self.processes.pop(ji).join()
-
-    def fill_queue(self) -> None:
-        """Add tile reads to the queue until the max number of workers is reached."""
-        while len(self.enqueued) < self.num_workers and len(self.remaining_reads) > 0:
-            next_ji = self.remaining_reads.pop(0)
-            process = multiprocessing.Process(
-                target=get_tile,
-                args=(
-                    self.queue,
-                    next_ji,
-                    self.read_tile_size,
-                    self.reader.path,
-                ),
-                daemon=True,
-            )
-            process.start()
-            self.processes[next_ji] = process
-            self.enqueued.add(next_ji)
-
-    def update_read_pbar(self) -> None:
-        """Update the read progress bar."""
-        if self.read_pbar is not None:
-            self.read_pbar.update()
-
-    def pop_next_read_tile(self) -> Optional[np.ndarray]:
-        """Remove the next tile from the reordering dict.
-
-        Returns:
-            Optional[np.ndarray]:
-                The next tile from the reordering dict if available.
-                If an intermediate is being used, or the tile is not
-                in the reordering dict, this will be None.
-        """
-        read_ji = (self.read_j, self.read_i)
-        if read_ji in self.reordering_dict:
-            self.enqueued.remove(read_ji)
-            tile = self.reordering_dict.pop(read_ji)
-
-            # If no intermediate is required, return the tile
-            if not self.intermediate:
-                if tile is None:
-                    raise Exception(f"Tile {read_ji} is None")
-                self.read_i += 1
-                self.update_read_pbar()
-                self.yield_i += 1
-                return tile
-
-            # Otherwise, write the tile to the intermediate
-            intermediate_write_index = tile_slices(
-                index=read_ji,
-                shape=self.read_tile_size,
-            )
-            self.intermediate[intermediate_write_index] = tile
-            tile_status_index = tuple(
-                slice(max(0, floor(x.start / r)), ceil(x.stop / r))
-                for x, r in zip(intermediate_write_index, self.yield_tile_size)
-            )
-            self.tile_status[tile_status_index] = 1
-            self.read_i += 1
-            self.update_read_pbar()
-        return None
-
-    def close(self):
-        """Safely end any dependants (threads, processes, and files).
-
-        Close progress bars and join child processes. Terminate children
-        if they fail to join after one second.
-        """
-        if self.read_pbar is not None:
-            self.read_pbar.close()
-        # Join processes in parallel threads
-        if self.processes:
-            with ThreadPoolExecutor(len(self.processes)) as executor:
-                executor.map(lambda p: p.join(1), self.processes.values())
-            # Terminate any child processes if still alive
-            for process in self.processes.values():
-                if process.is_alive():
-                    process.terminate()
-
-    def __del__(self):
-        """Destructor."""
-        self.close()
+    def original_shape(self) -> Tuple[int, ...]:
+        """Return the original shape of the image."""
+        return self.shape
 
 
 class JP2Reader(Reader):
@@ -614,36 +320,77 @@ class TIFFReader(Reader):
         import tifffile
 
         super().__init__(path)
-        self.tiff = tifffile.TiffFile(str(path))
-        self.tiff_page = self.tiff.pages[0]
+        self._tiff = tifffile.TiffFile(str(path))
+        self._tiff_page = self._tiff.pages[0]
         self.microns_per_pixel = self._get_mpp()
-        self.array = self.tiff_page.asarray()
-        if self.tiff_page.axes == "SYX":
-            # Transpose SYX -> YXS
-            self.array = np.transpose(self.array, (0, 1, 2), (1, 2, 0))
-        self.shape = self.array.shape
-        self.dtype = self.array.dtype
-        self.axes = self.tiff.series[0].axes
-        self.is_tiled = self.tiff_page.is_tiled
+
+        # Handle reading as an xarray dataset
+        self._zarr_tiff_store = self._tiff.aszarr()
+        # Ensure that the zarr store is a group
+        if zarr.storage.contains_array(self._zarr_tiff_store):
+            # Create an in memory group
+            group = zarr.group()
+            # Add the array to the group under the key "0"
+            array = zarr.open(self._zarr_tiff_store, mode="r")
+            group[0] = array
+            # Copy attrs over to the group member (a zarr bug?)
+            if "_ARRAY_DIMENSIONS" not in array.attrs:
+                raise ValueError(
+                    f"No _ARRAY_DIMENSIONS found in {self._zarr_tiff_store}."
+                )
+            for key, value in array.attrs.items():
+                group[0].attrs[key] = value
+            self._zarr_tiff_store = group.store
+        # Open the store as an xarray dataset
+        self._zarr = zarr.open(self._zarr_tiff_store, mode="r")
+        self._dataset = xr.open_zarr(self._zarr_tiff_store, consolidated=False)
+        # Copy over the dtype of the dataset (xarray bug?)
+        for key, array in self._zarr.items():
+            self._dataset[key] = self._dataset[key].astype(array.dtype)
+        # Rename S to C
+        if "S" in self._dataset["0"].dims:
+            self._dataset["0"] = self._dataset["0"].rename({"S": "C"})
+        # Normalise axes to TZYXC
+        self._tzyxc_dataset = self._dataset.copy()
+        self._tzyxc_dataset["0"] = self._tzyxc_dataset["0"].expand_dims(
+            dim=[a for a in "TCZYX" if a not in self._tzyxc_dataset["0"].dims],
+        )
+        self._tzyxc_dataset["0"] = self._tzyxc_dataset["0"].transpose(
+            "T", "Z", "Y", "X", "C"
+        )
+
+        # Set default time and depth
+        self.default_t = 0
+        self.default_z = 0
+
+        # Set standard Reader attributes
+        self.shape = self._dataset["0"].shape
+        self.dtype = self._dataset["0"].dtype
+        self.is_tiled = self._tiff_page.is_tiled
         self.tile_shape = None
         self.mosaic_shape = None
         self.mosaic_byte_offsets = None
         self.mosaic_byte_counts = None
+        # Read tile shape if tiled
         if self.is_tiled:
-            self.tile_shape = (self.tiff_page.tilelength, self.tiff_page.tilewidth)
+            self.tile_shape = (self._tiff_page.tilelength, self._tiff_page.tilewidth)
             self.mosaic_shape = mosaic_shape(
-                array_shape=self.shape, tile_shape=self.tile_shape
+                array_shape=self._tiff_page.shape, tile_shape=self.tile_shape
             )
-            self.mosaic_byte_offsets = np.array(self.tiff_page.dataoffsets).reshape(
+            self.mosaic_byte_offsets = np.array(self._tiff_page.dataoffsets).reshape(
                 self.mosaic_shape
             )
-            self.mosaic_byte_counts = np.array(self.tiff_page.databytecounts).reshape(
+            self.mosaic_byte_counts = np.array(self._tiff_page.databytecounts).reshape(
                 self.mosaic_shape
             )
-        self.jpeg_tables = self.tiff_page.jpegtables
-        self.color_space: ColorSpace = ColorSpace.from_tiff(self.tiff_page.photometric)
-        self.codec: Codec = Codec.from_tiff(self.tiff_page.compression)
+        self.jpeg_tables = self._tiff_page.jpegtables
+        self.color_space: ColorSpace = ColorSpace.from_tiff(self._tiff_page.photometric)
+        self.codec: Codec = Codec.from_tiff(self._tiff_page.compression)
         self.compression_level = None  # To be filled in if known later
+
+    @property
+    def original_shape(self) -> Tuple[int, ...]:
+        return self._tiff_page.shape
 
     def _get_mpp(self) -> Optional[Tuple[float, float]]:
         """Get the microns per pixel for the image.
@@ -656,7 +403,7 @@ class TIFFReader(Reader):
                 If the resolution is not available, this will be None.
         """
         try:
-            tags = self.tiff_page.tags
+            tags = self._tiff_page.tags
             y_resolution = tags["YResolution"].value[0] / tags["YResolution"].value[1]
             x_resolution = tags["XResolution"].value[0] / tags["XResolution"].value[1]
             resolution_units = tags["ResolutionUnit"].value
@@ -679,20 +426,66 @@ class TIFFReader(Reader):
             np.ndarray:
                 The tile at index.
         """
+        if self.tile_shape is None:
+            raise ValueError("Image is not tiled.")
         flat_index = index[0] * self.tile_shape[1] + index[1]
-        fh = self.tiff.filehandle
+        fh = self._tiff.filehandle
         _ = fh.seek(self.mosaic_byte_offsets[index])
         data = fh.read(self.mosaic_byte_counts[index])
         if not decode:
             return data
-        tile, _, _ = self.tiff_page.decode(
-            data, flat_index, jpegtables=self.tiff_page.jpegtables
+        tile, _, shape = self._tiff_page.decode(
+            data, flat_index, jpegtables=self._tiff_page.jpegtables
         )
-        return tile
+        # tile may be None e.g. with NDPI blank tiles
+        if tile is None:
+            tile = np.zeros(shape, dtype=self.dtype)
+        return tile[0]
 
     def __getitem__(self, index: Tuple[Union[slice, int]]) -> np.ndarray:
         """Get pixel data at index."""
-        return self.array[index]
+        index = index if isinstance(index, tuple) else (index,)
+        index = (self.default_t, self.default_z) + index
+        return self._tzyxc_dataset["0"][index].as_numpy().data
+
+    def thumbnail(self, shape: Tuple[int, ...], approx_ok: bool = False) -> np.ndarray:
+        """Get a thumbnail of the image.
+
+        Uses xarray/dask block map to get the thumbnail.
+
+        Args:
+            shape (Tuple[int, ...]):
+                The shape of the thumbnail to get.
+            approx_ok (bool, optional):
+                Whether to use an approximate thumbnail. Defaults to False.
+
+        Returns:
+            np.ndarray:
+                The thumbnail of the image.
+        """
+        yxc = self._tzyxc_dataset["0"][self.default_t, self.default_z]
+        (
+            _,
+            _,
+            downsample,
+        ) = self._find_thumbnail_downsample(
+            shape,
+            yxc.shape[:2],
+            yxc.data.chunksize[:2],
+        )
+        thumbnail = (  # noqa: ECE001
+            yxc.coarsen(X=downsample, boundary="trim")
+            .mean()
+            .coarsen(Y=downsample, boundary="trim")
+            .mean()
+            .astype("u1")
+            .as_numpy()
+        )
+        return (
+            thumbnail.data
+            if approx_ok
+            else resize_array(thumbnail.data, shape, "bicubic")
+        )
 
 
 class DICOMWSIReader(Reader):
